requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.9
munch==2.5.0
numpy==1.18.4
Pillow==7.1.2
pretrainedmodels==0.7.4
pycosat==0.6.3
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
requests==2.23.0
ruamel-yaml==0.15.87
six==1.14.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.43.0
urllib3==1.25.8
